split adjudicator files by purpose
