+++ src/lib/filter/adjudicators/bindingAdjudicators.test.ts
@@ -5,47 +5,47 @@ import { expect, describe, it } from "@jest/globals";
 import { KubernetesObject } from "kubernetes-fluent-client";
 import { Binding, DeepPartial, ValidateActionResponse } from "../../types";
 import { Event } from "../../enums";
+import { bindsToNamespace, misboundDeleteWithDeletionTimestamp, misboundNamespace } from "./postCollection";
+import { defaultBinding, defaultFilters, defaultKubernetesObject } from "./defaultTestObjects";
 import {
-  bindsToNamespace,
-  carriedAnnotations,
-  carriedLabels,
-  carriedName,
-  carriedNamespace,
-  carriesAnnotations,
-  carriesLabels,
-  carriesName,
-  carriesNamespace,
+  definesDeletionTimestamp,
+  ignoresDeletionTimestamp,
+  definedName,
+  definesName,
+  ignoresName,
+  definedNameRegex,
+  definesNameRegex,
+  definedNamespaces,
+  definesNamespaces,
+  definedNamespaceRegexes,
+  definesNamespaceRegexes,
   definedAnnotations,
+  definesAnnotations,
+  definedLabels,
+  definesLabels,
+  definedEvent,
+  definesDelete,
+  definedGroup,
+  definesGroup,
+  definedVersion,
+  definesVersion,
+  definedKind,
+  definesKind,
+  definedCategory,
   definedCallback,
   definedCallbackName,
-  definedCategory,
-  definedEvent,
-  definedGroup,
-  definedKind,
-  definedLabels,
-  definedName,
-  definedNameRegex,
-  definedNamespaceRegexes,
-  definedNamespaces,
-  definedVersion,
-  definesAnnotations,
-  definesDelete,
-  definesDeletionTimestamp,
-  definesGroup,
-  definesKind,
-  definesLabels,
-  definesName,
-  definesNameRegex,
-  definesNamespaceRegexes,
-  definesNamespaces,
-  definesVersion,
-  ignoresDeletionTimestamp,
-  ignoresName,
-  misboundDeleteWithDeletionTimestamp,
-  misboundNamespace,
+} from "./binding";
+import {
+  carriedName,
+  carriesName,
   missingName,
-} from "./adjudicators";
-import { defaultBinding, defaultFilters, defaultKubernetesObject } from "./defaultTestObjects";
+  carriedNamespace,
+  carriesNamespace,
+  carriedAnnotations,
+  carriesAnnotations,
+  carriedLabels,
+  carriesLabels,
+} from "./kubernetesObject";
 
 describe("definesDeletionTimestamp", () => {
   //[ Binding, result ]
