+++ src/lib/filter/adjudicators/requestAdjudicators.test.ts
@@ -5,7 +5,7 @@ import { expect, describe, it } from "@jest/globals";
 import { Operation } from "../../enums";
 import { AdmissionRequest } from "../../types";
 import { defaultAdmissionRequest } from "./defaultTestObjects";
-import { declaredUid, declaredKind, declaredVersion, declaredGroup, declaredOperation } from "./adjudicators";
+import { declaredUid, declaredKind, declaredVersion, declaredGroup, declaredOperation } from "./admissionRequest";
 
 describe("declaredUid", () => {
   //[ AdmissionRequest, result ]
