+++ src/lib/filter/filter.ts
@@ -5,25 +5,7 @@ import { AdmissionRequest, Binding } from "../types";
 import { Operation } from "../enums";
 import { KubernetesObject } from "kubernetes-fluent-client";
 import {
-  carriedAnnotations,
-  carriedLabels,
-  carriedName,
-  carriedNamespace,
   carriesIgnoredNamespace,
-  declaredGroup,
-  declaredKind,
-  declaredOperation,
-  declaredVersion,
-  definedAnnotations,
-  definedEvent,
-  definedGroup,
-  definedKind,
-  definedLabels,
-  definedName,
-  definedNameRegex,
-  definedNamespaceRegexes,
-  definedNamespaces,
-  definedVersion,
   misboundDeleteWithDeletionTimestamp,
   misboundNamespace,
   mismatchedAnnotations,
@@ -40,7 +22,21 @@ import {
   missingCarriableNamespace,
   unbindableNamespaces,
   uncarryableNamespace,
-} from "./adjudicators/adjudicators";
+} from "./adjudicators/postCollection";
+import { declaredOperation, declaredGroup, declaredVersion, declaredKind } from "./adjudicators/admissionRequest";
+import {
+  definedEvent,
+  definedName,
+  definedGroup,
+  definedVersion,
+  definedKind,
+  definedNamespaces,
+  definedLabels,
+  definedAnnotations,
+  definedNamespaceRegexes,
+  definedNameRegex,
+} from "./adjudicators/binding";
+import { carriedName, carriedNamespace, carriedLabels, carriedAnnotations } from "./adjudicators/kubernetesObject";
 
 type AdjudicationResult = string | null;
 type Adjudicator = () => AdjudicationResult;
